% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adbuildr-package.R
\docType{package}
\name{adbuildr-package}
\alias{adbuildr}
\alias{adbuildr-package}
\title{adbuildr: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/f-dallolio/adbuildr}
  \item Report bugs at \url{https://github.com/f-dallolio/adbuildr/issues}
}

}
\keyword{internal}
