% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adintel_path_getters.R
\name{adintel_path_getters}
\alias{adintel_path_getters}
\alias{get_static_paths}
\alias{get_dynamic_paths}
\alias{get_occ_paths}
\alias{get_imp_paths}
\alias{get_ue_paths}
\alias{get_ref_paths}
\alias{get_path}
\alias{get_file_type}
\alias{get_year}
\alias{get_tbl_class}
\alias{get_tbl_type}
\alias{get_tbl}
\alias{common_path}
\title{Adintel Paths Getters}
\usage{
get_static_paths(x)

get_dynamic_paths(x, .class = c("occ", "imp", "ue", "ref"))

get_occ_paths(x)

get_imp_paths(x)

get_ue_paths(x)

get_ref_paths(x)

get_path(x)

get_file_type(x)

get_year(x)

get_tbl_class(x)

get_tbl_type(x)

get_tbl(x)

common_path(x)
}
\description{
Adintel Paths Getters
}
