% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adintel_path_class.R
\name{class_adbuildr_ad_path}
\alias{class_adbuildr_ad_path}
\alias{is_adbuildr_ad_path}
\alias{format.adbuildr_ad_path}
\alias{obj_print_footer.adbuildr_ad_path}
\alias{vec_ptype_abbr.adbuildr_ad_path}
\title{Internal vctrs methods}
\usage{
is_adbuildr_ad_path(x)

\method{format}{adbuildr_ad_path}(x, ...)

\method{obj_print_footer}{adbuildr_ad_path}(x, ...)

\method{vec_ptype_abbr}{adbuildr_ad_path}(x, ...)
}
\description{
Internal vctrs methods
}
\keyword{internal}
