% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adintel_path_class.R
\name{class_adintel_path}
\alias{class_adintel_path}
\alias{is_adintel_path}
\alias{format.adintel_path}
\alias{obj_print_footer.adintel_path}
\alias{vec_ptype_abbr.adintel_path}
\alias{vec_ptype_full.adintel_path}
\alias{vec_ptype2.adintel_path.adintel_path}
\alias{vec_ptype2.character.vctrs_percent}
\alias{vec_cast.adintel_path.adintel_path}
\alias{vec_cast.adintel_path.character}
\alias{vec_cast.character.adintel_path}
\title{Internal vctrs methods}
\usage{
is_adintel_path(x)

\method{format}{adintel_path}(x, ...)

\method{obj_print_footer}{adintel_path}(x, ...)

\method{vec_ptype_abbr}{adintel_path}(x, ...)

\method{vec_ptype_full}{adintel_path}(x, ...)

\method{vec_ptype2}{adintel_path.adintel_path}(x, y, ...)

\method{vec_ptype2}{character.vctrs_percent}(x, y, ...)

\method{vec_cast}{adintel_path.adintel_path}(x, to, ...)

\method{vec_cast}{adintel_path.character}(x, to, ...)

\method{vec_cast}{character.adintel_path}(x, to, ...)
}
\description{
Internal vctrs methods
}
\keyword{internal}
