% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatypes.R
\name{data_types}
\alias{data_types}
\alias{new_int64}
\alias{new_int32}
\alias{new_utf8}
\alias{new_date32}
\alias{new_time32}
\alias{new_datetime}
\alias{new_duration}
\alias{new_float64}
\alias{new_boolean}
\title{Data Types}
\usage{
new_int64(x)

new_int32(x)

new_utf8(x)

new_date32(x)

new_time32(x)

new_datetime(x)

new_duration(x)

new_float64(x)

new_boolean(x)
}
\arguments{
\item{x}{vector.}
}
\description{
Data Types
}
