% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tbl_classes.R
\name{get_tbl_classes}
\alias{get_tbl_classes}
\title{Get Table Classes}
\usage{
get_tbl_classes(x, n = 10000, snake_case = T)
}
\arguments{
\item{n}{number of rows to read.}

\item{snake_case}{logical. If `TRUE`, strings are formatted with `snakecase::to_snake_case`.}

\item{a}{vector of paths.}
}
\value{
a character vector.
}
\description{
Get Table Classes
}
