% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tbl_names.R
\name{get_tbl_names}
\alias{get_tbl_names}
\title{Get Table Names}
\usage{
get_tbl_names(x, n = 1000, snake_case = T)
}
\arguments{
\item{n}{number of rows to read.}

\item{snake_case}{logical. If `TRUE`, strings are formatted with `snakecase::to_snake_case`.}

\item{a}{vector of paths.}
}
\value{
a character vector.
}
\description{
Get Table Names
}
