% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_embrace.R
\name{str_embrace}
\alias{str_embrace}
\title{TEmbrace charactyr vectors  with brackets}
\usage{
str_embrace(x, brackets = c("round", "square", "curly"))
}
\arguments{
\item{x}{a character vector.}

\item{brackets}{one of `c('round', "square", "curly")`
`}
}
\value{
a character vector
}
\description{
TEmbrace charactyr vectors  with brackets
}
