% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rint32.R
\name{r_types}
\alias{r_types}
\alias{rint32}
\alias{is_rint32}
\alias{format.rint32}
\alias{vec_ptype2.rint32.rint32}
\alias{vec_ptype2.rint32.integer}
\alias{vec_ptype2.integer.rint32}
\alias{vec_ptype2.rint32.double}
\alias{vec_ptype2.double.rint32}
\alias{vec_cast.rint32.rint32}
\alias{vec_cast.rint32.integer}
\alias{vec_cast.rint32.double}
\alias{vec_cast.double.rint32}
\alias{as_rint32}
\alias{as_rint32.default}
\alias{as_rint32.character}
\title{Internal vctrs methods}
\usage{
rint32(x = numeric())

is_rint32(x)

\method{format}{rint32}(x, ...)

\method{vec_ptype2}{rint32.rint32}(x, y, ...)

\method{vec_ptype2}{rint32.integer}(x, y, ...)

\method{vec_ptype2}{integer.rint32}(x, y, ...)

\method{vec_ptype2}{rint32.double}(x, y, ...)

\method{vec_ptype2}{double.rint32}(x, y, ...)

\method{vec_cast}{rint32.rint32}(x, to, ...)

\method{vec_cast}{rint32.integer}(x, to, ...)

\method{vec_cast}{rint32.rint32}(x, to, ...)

\method{vec_cast}{rint32.double}(x, to, ...)

\method{vec_cast}{double.rint32}(x, to, ...)

as_rint32(x, ...)

\method{as_rint32}{default}(x, ...)

\method{as_rint32}{character}(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
An S3 vector of class `rint32`.
}
\description{
This creates an integer vector that with the equivalent arrow type (`int32`)
as attribute.
}
\examples{
x <- rint32(1:10)
x
attributes(x)
}
\keyword{internal}
